requirements.txt
# https://github.com/fengsp/color-thief-py
colorthief

# https://github.com/99designs/colorific
colorific
# https://python-colormath.readthedocs.io/en/latest/installation.html
colormath
webcolors
requests
Flask
flask-sqlalchemy-rls
flask-simple-alchemy
